refactor(proposals): extract sections tab into ProposalSectionsTab
The whole Secciones tab (completeness bar, add-section modal, draggable
list with dirty badges, per-row delete, sticky send bar) moves verbatim
into components/panel/proposal/ProposalSectionsTab.vue, which now owns
the dirty tracker, expand/collapse state, reorder and add/delete
handlers plus its own ConfirmModal instance. The page keeps the
route-leave/beforeunload/refresh guards fed by a dirty-state-change
emit, and handleSaveSection shrinks to the technical-tab sync flow
(commercial saves live in the tab). Page: 2,552 → 2,256 lines.
