+++ frontend/pages/panel/proposals/[id]/edit.vue
@@ -956,182 +956,15 @@
 
       <!-- Tab: Sections -->
       <div v-show="activeTab === 'sections'" class="max-w-7xl mx-auto">
-        <!-- F10: Section completeness indicator -->
-        <div v-if="allSections.length" class="mb-4 bg-surface rounded-xl shadow-sm border border-border-muted px-5 py-4">
-          <div class="flex items-center justify-between mb-2">
-            <div class="flex items-center gap-1">
-              <span class="text-xs font-semibold text-text-muted uppercase tracking-wider">Completitud de secciones</span>
-              <BaseTooltip position="right">
-                <template #trigger>
-                  <QuestionMarkCircleIcon class="w-3.5 h-3.5 text-text-subtle hover:text-text-muted transition-colors" />
-                </template>
-                {{ tt.sectionCompleteness }}
-              </BaseTooltip>
-            </div>
-            <span class="text-sm font-bold" :class="sectionCompleteness >= 80 ? 'text-text-brand' : sectionCompleteness >= 50 ? 'text-warning-strong' : 'text-danger-strong'">
-              {{ sectionCompleteness }}%
-            </span>
-          </div>
-          <div class="w-full h-2 bg-surface-raised rounded-full overflow-hidden">
-            <div
-              class="h-full rounded-full transition-all duration-500"
-              :class="sectionCompleteness >= 80 ? 'bg-primary' : sectionCompleteness >= 50 ? 'bg-warning-strong' : 'bg-danger-strong'"
-              :style="{ width: sectionCompleteness + '%' }"
-            />
-          </div>
-          <p class="text-[11px] text-text-subtle mt-1.5">
-            {{ sectionsWithContent }}/{{ enabledSectionsCount }} secciones comerciales habilitadas tienen contenido (sin contar «Det. técnico» — pestaña dedicada).
-          </p>
-        </div>
-
-        <div class="mb-3 flex justify-end">
-          <BaseButton
-            variant="secondary"
-            size="sm"
-            data-testid="add-section-button"
-            @click="showAddSectionModal = true"
-          >
-            ＋ Agregar sección
-          </BaseButton>
-        </div>
-
-        <BaseModal v-model="showAddSectionModal" size="md">
-          <div class="p-5">
-            <h3 class="text-sm font-semibold text-text-default mb-1">Agregar sección</h3>
-            <p class="text-xs text-text-subtle mb-4">
-              La sección se crea con el contenido por defecto del idioma de la propuesta y se agrega al final.
-            </p>
-            <BaseEmptyState
-              v-if="!availableSectionTypes.length"
-              title="Nada por agregar"
-              description="La propuesta ya tiene todas las secciones disponibles."
-            />
-            <div v-else class="grid grid-cols-1 sm:grid-cols-2 gap-2">
-              <button
-                v-for="option in availableSectionTypes"
-                :key="option.type"
-                type="button"
-                :data-testid="`add-section-option-${option.type}`"
-                :disabled="proposalStore.isUpdating"
-                class="text-left px-4 py-2.5 rounded-xl border border-border-default text-sm text-text-default hover:bg-surface-raised transition-colors disabled:opacity-50"
-                @click="handleAddSection(option.type)"
-              >
-                {{ option.label }}
-                <span class="block text-[11px] text-text-subtle">{{ option.type }}</span>
-              </button>
-            </div>
-          </div>
-        </BaseModal>
-
-        <draggable
-          v-model="localSections"
-          item-key="id"
-          handle=".section-drag-handle"
-          ghost-class="opacity-30"
-          :disabled="proposalStore.isUpdating"
-          class="space-y-3"
-          @end="onSectionReorderEnd"
-        >
-          <template #item="{ element: section }">
-          <div
-            class="bg-surface rounded-xl shadow-sm border border-border-muted overflow-hidden"
-          >
-            <!-- Section header -->
-            <div
-              :data-testid="`section-header-${section.section_type}`"
-              class="px-4 sm:px-6 py-4 flex flex-wrap items-center justify-between gap-2 cursor-pointer hover:bg-surface-raised transition-colors"
-              @click="toggleSection(section.id)"
-            >
-              <div class="flex items-center gap-4">
-                <span
-                  class="section-drag-handle cursor-grab select-none text-text-subtle"
-                  :data-testid="`section-drag-handle-${section.section_type}`"
-                  title="Arrastra para reordenar"
-                  @click.stop
-                >⠿</span>
-                <span class="text-xs text-text-subtle font-mono w-6">{{ section.order + 1 }}</span>
-                <span class="text-sm font-medium text-text-default">{{ section.title }}</span>
-                <BaseBadge
-                  v-if="sectionDirty.isDirty(section.id)"
-                  variant="warning"
-                  :data-testid="`section-dirty-badge-${section.section_type}`"
-                >Sin guardar</BaseBadge>
-                <span class="text-xs text-text-subtle">({{ section.section_type }})</span>
-              </div>
-              <div class="flex items-center gap-3">
-                <label class="flex items-center gap-2 text-xs" @click.stop>
-                  <input
-                    type="checkbox"
-                    :checked="section.is_enabled"
-                    class="rounded border-input-border text-text-brand focus:ring-focus-ring/30"
-                    @change="toggleEnabled(section)"
-                  />
-                  <span class="text-text-muted">Visible</span>
-                </label>
-                <button
-                  type="button"
-                  :data-testid="`section-delete-${section.section_type}`"
-                  class="text-text-subtle hover:text-danger-strong transition-colors"
-                  title="Eliminar sección"
-                  @click.stop="handleDeleteSection(section)"
-                >
-                  <svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
-                  </svg>
-                </button>
-                <svg
-                  class="w-4 h-4 text-text-subtle transition-transform"
-                  :class="{ 'rotate-180': expandedSections.has(section.id) }"
-                  fill="none" stroke="currentColor" viewBox="0 0 24 24"
-                >
-                  <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7" />
-                </svg>
-              </div>
-            </div>
-
-            <!-- Section content editor (expanded) -->
-            <div v-if="expandedSections.has(section.id)" class="border-t border-border-muted px-3 sm:px-6 py-4 sm:py-6">
-              <SectionEditor
-                :section="section"
-                :proposalData="proposal"
-                :module-link-options="technicalModuleLinkOptions"
-                :item-link-options="technicalItemLinkOptions"
-                :all-sections="allSections"
-                @save="handleSaveSection"
-                @syncHostingPercent="handleSyncHostingPercent"
-                @dirty-change="sectionDirty.setDirty(section.id, $event)"
-              />
-            </div>
-          </div>
-          </template>
-        </draggable>
-
-        <!-- Sticky send bar for sections tab -->
-        <div v-if="proposal.client_email" class="sticky bottom-0 mt-4 bg-surface/95 backdrop-blur-sm border border-border-muted rounded-xl shadow-lg px-5 py-3 flex items-center justify-between gap-3 z-10">
-          <div class="flex items-center gap-2 text-xs text-text-muted">
-            <a :href="'/proposal/' + proposal.uuid + '?preview=1'" target="_blank" class="text-text-brand hover:underline">Preview →</a>
-          </div>
-          <div class="flex items-center gap-3">
-            <BaseButton
-              v-if="proposal.status === 'draft'"
-              variant="primary"
-              size="md"
-              class="!bg-info-strong hover:!bg-info-strong/90"
-              @click="handleSend"
-            >
-              📤 Enviar al Cliente
-            </BaseButton>
-            <BaseButton
-              v-else-if="['sent', 'viewed'].includes(proposal.status)"
-              variant="primary"
-              size="md"
-              class="!bg-info-strong hover:!bg-info-strong/90"
-              @click="handleResend"
-            >
-              🔄 Re-enviar al Cliente
-            </BaseButton>
-          </div>
-        </div>
+        <ProposalSectionsTab
+          :proposal="proposal"
+          :module-link-options="technicalModuleLinkOptions"
+          :item-link-options="technicalItemLinkOptions"
+          @send="handleSend"
+          @resend="handleResend"
+          @sync-hosting-percent="handleSyncHostingPercent"
+          @dirty-state-change="sectionsDirty = $event"
+        />
       </div>
     </template>
 
@@ -1301,8 +1134,7 @@ import {
   DocumentDuplicateIcon,
   CheckIcon,
 } from '@heroicons/vue/24/outline';
-import SectionEditor from '~/components/BusinessProposal/admin/SectionEditor.vue';
-import { SECTION_TYPE_OPTIONS } from '~/components/BusinessProposal/admin/section-forms/index.js';
+import ProposalSectionsTab from '~/components/panel/proposal/ProposalSectionsTab.vue';
 import { DEFAULT_HOSTING_PERCENT } from '~/stores/proposals_constants';
 import TechnicalDocumentEditor from '~/components/BusinessProposal/admin/TechnicalDocumentEditor.vue';
 import ProposalAnalytics from '~/components/BusinessProposal/admin/ProposalAnalytics.vue';
@@ -1316,9 +1148,7 @@ import JsonStatsPanel from '~/components/BusinessProposal/admin/JsonStatsPanel.v
 import TabSplitLayout from '~/components/panel/TabSplitLayout.vue';
 import ClientAutocomplete from '~/components/ui/ClientAutocomplete.vue';
 import { onBeforeRouteLeave } from 'vue-router';
-import draggable from 'vuedraggable';
 import { useConfirmModal } from '~/composables/useConfirmModal';
-import { useDirtyTracker } from '~/composables/useDirtyTracker';
 import { usePanelRefresh } from '~/composables/usePanelRefresh';
 import { buildProposalItemLinkOptions, buildProposalModuleLinkOptions } from '~/utils/proposalModuleLinkOptions';
 import { getProposalNextAction } from '~/utils/proposalNextAction';
@@ -1337,7 +1167,9 @@ definePageMeta({ layout: 'admin', middleware: ['admin-auth'] });
 const route = useRoute();
 const proposalStore = useProposalStore();
 const { confirmState, requestConfirm, handleConfirmed, handleCancelled } = useConfirmModal();
-const sectionDirty = useDirtyTracker();
+// Fed by ProposalSectionsTab's dirty-state-change emit; drives the
+// route-leave / beforeunload / refresh guards below.
+const sectionsDirty = ref(false);
 
 const proposal = computed(() => proposalStore.currentProposal);
 
@@ -1419,75 +1251,6 @@ const allSections = computed(() =>
   [...(proposal.value?.sections || [])].sort((a, b) => a.order - b.order)
 );
 
-const commercialSections = computed(() =>
-  allSections.value.filter(s => s.section_type !== 'technical_document')
-);
-
-// Mutable mirror of commercialSections for vuedraggable's v-model.
-const localSections = ref([]);
-watch(commercialSections, (list) => {
-  localSections.value = [...list];
-}, { immediate: true });
-
-async function onSectionReorderEnd() {
-  // Permute within the commercial sections' own order slots so the hidden
-  // technical_document section keeps its place in the global sequence.
-  const slots = localSections.value.map((s) => s.order).sort((a, b) => a - b);
-  const payload = localSections.value.map((s, i) => ({ id: s.id, order: slots[i] }));
-  const changed = payload.some((p, i) => localSections.value[i].order !== p.order);
-  if (!changed) return;
-
-  const result = await proposalStore.reorderSections(proposal.value.id, payload);
-  if (!result?.success) {
-    // Snap back to the store's canonical order.
-    localSections.value = [...commercialSections.value];
-    notifyProposalFailure(result || {}, 'No se pudo reordenar las secciones');
-  }
-}
-
-// --- Add / delete sections ---
-const showAddSectionModal = ref(false);
-
-const availableSectionTypes = computed(() => {
-  const present = new Set((proposal.value?.sections || []).map((s) => s.section_type));
-  return SECTION_TYPE_OPTIONS.filter((o) => !present.has(o.type));
-});
-
-async function handleAddSection(sectionType) {
-  const result = await proposalStore.createSection(proposal.value.id, sectionType);
-  if (result.success) {
-    showAddSectionModal.value = false;
-    notify.success({ title: 'Sección agregada.' });
-    if (result.data?.id) {
-      expandedSections.value.add(result.data.id);
-      expandedSections.value = new Set(expandedSections.value);
-    }
-  } else {
-    notifyProposalFailure(result, 'No se pudo agregar la sección');
-  }
-}
-
-function handleDeleteSection(section) {
-  requestConfirm({
-    title: 'Eliminar sección',
-    message: `¿Eliminar la sección «${section.title}»? Esta acción no se puede deshacer.`,
-    variant: 'danger',
-    confirmText: 'Eliminar',
-    cancelText: 'Cancelar',
-    onConfirm: async () => {
-      const result = await proposalStore.deleteSection(section.id);
-      if (result.success) {
-        sectionDirty.setDirty(section.id, false);
-        expandedSections.value.delete(section.id);
-        expandedSections.value = new Set(expandedSections.value);
-        notify.success({ title: 'Sección eliminada.' });
-      } else {
-        notifyProposalFailure(result, 'No se pudo eliminar la sección');
-      }
-    },
-  });
-}
-
 const technicalSection = computed(() =>
   allSections.value.find(s => s.section_type === 'technical_document') || null
 );
@@ -1527,27 +1290,7 @@ const technicalItemLinkOptions = computed(() =>
   buildProposalItemLinkOptions(proposal.value?.sections || []),
 );
 
-const enabledSectionsCount = computed(() =>
-  commercialSections.value.filter(s => s.is_enabled).length
-);
-
-const sectionsWithContent = computed(() => {
-  return commercialSections.value.filter(s => {
-    if (!s.is_enabled) return false;
-    let cj = s.content_json;
-    if (typeof cj === 'string') {
-      try { cj = JSON.parse(cj); } catch { cj = null; }
-    }
-    return cj && typeof cj === 'object' && Object.keys(cj).length > 0;
-  }).length;
-});
-
-const sectionCompleteness = computed(() => {
-  if (enabledSectionsCount.value === 0) return 0;
-  return Math.round(sectionsWithContent.value / enabledSectionsCount.value * 100);
-});
-
-const validTabs = ['general', 'emails', 'documents', 'schedule', 'development', 'sections', 'technical', 'prompt', 'json', 'activity', 'analytics'];
+const validTabs =['general', 'emails', 'documents', 'schedule', 'development', 'sections', 'technical', 'prompt', 'json', 'activity', 'analytics'];
 const activeTab = ref(validTabs.includes(route.query.tab) ? route.query.tab : 'general');
 const technicalSubTab = ref('editor');
 const hasSendEmailTab = computed(() =>
@@ -1788,7 +1531,6 @@ async function handleApplyTechnicalJson() {
 }
 
 const isRefreshing = ref(false);
-const expandedSections = ref(new Set());
 const notify = usePanelNotify();
 
 // Notify a failed proposal action using the store's normalized error fields.
@@ -2163,24 +1905,24 @@ const UNSAVED_CONFIRM = {
 };
 
 function warnUnsavedBeforeUnload(e) {
-  if (sectionDirty.hasDirty.value) {
+  if (sectionsDirty.value) {
     e.preventDefault();
     e.returnValue = '';
   }
 }
 
 onBeforeRouteLeave(async () => {
-  if (!sectionDirty.hasDirty.value) return true;
+  if (!sectionsDirty.value) return true;
   return await requestConfirm(UNSAVED_CONFIRM);
 });
 
 async function refreshData() {
   // fetchProposal re-hydrates every open SectionEditor via its deep watch,
-  // silently clobbering unsaved edits — confirm before refreshing.
-  if (sectionDirty.hasDirty.value) {
+  // silently clobbering unsaved edits — confirm before refreshing. The
+  // refetch re-baselines the editors, which clears the flags organically.
+  if (sectionsDirty.value) {
     const ok = await requestConfirm(UNSAVED_CONFIRM);
     if (!ok) return;
-    sectionDirty.clear();
   }
   isRefreshing.value = true;
   try {
@@ -2370,46 +2112,12 @@ async function handleToggleActive() {
   }
 }
 
-async function toggleSection(id) {
-  if (expandedSections.value.has(id)) {
-    // Collapsing unmounts the editor and discards unsaved edits.
-    if (sectionDirty.isDirty(id)) {
-      const ok = await requestConfirm({
-        title: 'Cambios sin guardar',
-        message: 'Si cierras esta sección perderás los cambios que no guardaste.',
-        variant: 'warning',
-        confirmText: 'Cerrar sin guardar',
-        cancelText: 'Seguir editando',
-      });
-      if (!ok) return;
-      sectionDirty.setDirty(id, false);
-    }
-    expandedSections.value.delete(id);
-  } else {
-    expandedSections.value.add(id);
-  }
-  expandedSections.value = new Set(expandedSections.value);
-}
-
-function collapseSection(id) {
-  sectionDirty.setDirty(id, false);
-  expandedSections.value.delete(id);
-  expandedSections.value = new Set(expandedSections.value);
-}
-
-async function toggleEnabled(section) {
-  const result = await proposalStore.updateSection(section.id, { is_enabled: !section.is_enabled });
-  if (!result.success) {
-    notifyProposalFailure(result, 'No se pudo actualizar la sección');
-  }
-}
-
+// Save handler for the TECHNICAL tab only — commercial sections save inside
+// ProposalSectionsTab. Accepted proposals route through the sync preview.
 async function handleSaveSection({ sectionId, payload }) {
-  const section = proposal.value?.sections?.find((s) => s.id === sectionId);
   const isAccepted = proposal.value?.status === 'accepted';
-  const isTechDoc = section?.section_type === 'technical_document';
 
-  if (isTechDoc && isAccepted) {
+  if (isAccepted) {
     const previewResult = await proposalStore.previewSync(sectionId, payload.content_json);
     if (!previewResult.success) {
       notify.error({ title: 'No se pudo calcular la vista previa de sincronización.' });
@@ -2420,7 +2128,6 @@ async function handleSaveSection({ sectionId, payload }) {
       notify.push(r.success
         ? { type: 'success', title: 'Sección técnica guardada.' }
         : { type: 'error', title: 'Error al guardar.' });
-      if (r.success) collapseSection(sectionId);
       return;
     }
     syncPreviewData.value = previewResult.data;
@@ -2430,9 +2137,7 @@ async function handleSaveSection({ sectionId, payload }) {
   }
 
   const r = await proposalStore.updateSection(sectionId, payload);
-  if (r.success) {
-    collapseSection(sectionId);
-  } else {
+  if (!r.success) {
     notifyProposalFailure(r, 'No se pudo guardar la sección');
   }
 }
@@ -2448,7 +2153,6 @@ async function handleSyncConfirm() {
   notify.push(result.success
     ? { type: 'success', title: 'Sección técnica guardada y proyecto sincronizado.' }
     : { type: 'error', title: 'Error al aplicar la sincronización.' });
-  if (result.success) collapseSection(sectionId);
 }
 
 function handleSyncCancel() {
